// client/dist/widgets/common/search/src/config.ts
export interface InitResultServiceListOption {
  configId?: string
  magicKey?: string
  isFromSuggestion?: boolean
}

export enum SearchResultView {
  ResultPanel = 'Result_Panel',
  OtherWidgets = 'Other_Widgets'
}

export enum ArrangementStyle {
  Style1 = 'Style1',
  Style2 = 'Style2',
  Style3 = 'Style3'
}

export enum SearchResultStyle {
  Classic = 'Classic',
  Compact = 'Compact'
}
// ...
export interface SuggestionItem {
  suggestionHtml: string | Element
  suggestion: string
  isRecentSearche?: boolean
  configId?: string
  isFromSuggestion?: boolean
  magicKey?: string
}

export interface Suggestion {
  suggestionItem: SuggestionItem[]
  layer: string
  icon?: any
}
// ...
export interface config extends SearchSuggestionConfig {
  datasourceConfig: SearchDataConfig[]
  hint: string
  searchResultView: SearchResultView
  arrangementStyle: ArrangementStyle
  searchResultStyle: SearchResultStyle
  resultMaxNumber: number
  isAutoSelectFirstResult: boolean
  // link
  linkParam?: IMLinkParam
}

export type IMConfig = ImmutableObject<config>